PBXBuildPhase abstract class
