+++ Sources/PBXFrameworksBuildPhase.swift
@@ -2,36 +2,7 @@ import Foundation
 import Unbox
 
 // This is the element for the framewrok link build phase.
-public class PBXFrameworksBuildPhase: ProjectElement {
-    
-    // MARK: - Properties
-    
-    /// Framework build phase files.
-    public var files: Set<String>
-    
-    /// Build phase build action mask.
-    public var buildActionMask: Int
-    
-    /// Build phase run only for deployment post processing.
-    public var runOnlyForDeploymentPostprocessing: UInt
-    
-    // MARK: - Init
-    
-    /// Initializes the frameworks build phase with its attributes.
-    ///
-    /// - Parameters:
-    ///   - reference: element reference.
-    ///   - files: frameworks build phase files.
-    ///   - runOnlyForDeploymentPostprocessing: run only for deployment pos processing value.
-    public init(reference: String,
-                files: Set<String>,
-                runOnlyForDeploymentPostprocessing: UInt,
-                buildActionMask: Int = 2147483647) {
-        self.files = files
-        self.runOnlyForDeploymentPostprocessing = runOnlyForDeploymentPostprocessing
-        self.buildActionMask = buildActionMask
-        super.init(reference: reference)
-    }
+public class PBXFrameworksBuildPhase: PBXBuildPhase {
 
     public static var isa: String = "PBXFrameworksBuildPhase"
 
@@ -40,14 +11,6 @@ public class PBXFrameworksBuildPhase: ProjectElement {
         return lhs.reference == rhs.reference &&
             lhs.files == rhs.files &&
             lhs.runOnlyForDeploymentPostprocessing == rhs.runOnlyForDeploymentPostprocessing
-    }
-    
-    public override init(reference: String, dictionary: [String: Any]) throws {
-        let unboxer = Unboxer(dictionary: dictionary)
-        self.files = try unboxer.unbox(key: "files")
-        self.runOnlyForDeploymentPostprocessing = try unboxer.unbox(key: "runOnlyForDeploymentPostprocessing")
-        self.buildActionMask = try unboxer.unbox(key: "buildActionMask")
-        try super.init(reference: reference, dictionary: dictionary)
     }
 }
