+++ Sources/PBXShellScriptBuildPhase.swift
@@ -2,15 +2,9 @@ import Foundation
 import Unbox
 
 // This is the element for the resources copy build phase.
-public class PBXShellScriptBuildPhase: ProjectElement {
+public class PBXShellScriptBuildPhase: PBXBuildPhase {
 
     // MARK: - Attributes
-
-    /// Files references.
-    public var files: Set<String>
-
-    /// Build action mask.
-    public var buildActionMask: Int
 
     /// Build phase name.
     public var name: String
@@ -20,9 +14,6 @@ public class PBXShellScriptBuildPhase: ProjectElement {
 
     /// Output paths
     public var outputPaths: Set<String>
-
-    /// Run only for deployment post processing attribute.
-    public var runOnlyForDeploymentPostprocessing: Int
 
     /// Path to the shell.
     public var shellPath: String
@@ -49,31 +40,25 @@ public class PBXShellScriptBuildPhase: ProjectElement {
                 outputPaths: Set<String>,
                 shellPath: String = "bin/sh",
                 shellScript: String?,
-                buildActionMask: Int = 2147483647,
-                runOnlyForDeploymentPostprocessing: Int = 0) {
-        self.files = files
+                buildActionMask: UInt = 2147483647,
+                runOnlyForDeploymentPostprocessing: UInt = 0) {
         self.name = name
         self.inputPaths = inputPaths
         self.outputPaths = outputPaths
         self.shellPath = shellPath
         self.shellScript = shellScript
-        self.buildActionMask = buildActionMask
-        self.runOnlyForDeploymentPostprocessing = runOnlyForDeploymentPostprocessing
-        super.init(reference: reference)
+        super.init(reference: reference, files: files, buildActionMask: buildActionMask, runOnlyForDeploymentPostprocessing: runOnlyForDeploymentPostprocessing)
     }
 
     public static var isa: String = "PBXShellScriptBuildPhase"
 
     public override init(reference: String, dictionary: [String: Any]) throws {
         let unboxer = Unboxer(dictionary: dictionary)
-        self.files = try unboxer.unbox(key: "files")
         self.name = try unboxer.unbox(key: "name")
         self.inputPaths = (unboxer.unbox(key: "inputPaths")) ?? []
         self.outputPaths = (unboxer.unbox(key: "outputPaths")) ?? []
         self.shellPath = try unboxer.unbox(key: "shellPath")
         self.shellScript = unboxer.unbox(key: "shellScript")
-        self.buildActionMask = try unboxer.unbox(key: "buildActionMask")
-        self.runOnlyForDeploymentPostprocessing = unboxer.unbox(key: "runOnlyForDeploymentPostprocessing") ?? 0
         try super.init(reference: reference, dictionary: dictionary)
     }
