+++ Tests/xcodeprojTests/PBXHeadersBuildPhaseSpec.swift
@@ -10,8 +10,8 @@ final class PBXHeadersBuildPhaseSpec: XCTestCase {
     override func setUp() {
         super.setUp()
         subject = PBXHeadersBuildPhase(reference: "ref",
-                                       buildActionMask: 0,
                                        files: Set(arrayLiteral: "333"),
+                                       buildActionMask: 0,
                                        runOnlyForDeploymentPostprocessing: 0)
     }
     
@@ -55,8 +55,8 @@ final class PBXHeadersBuildPhaseSpec: XCTestCase {
     
     func test_equals_returnsTheCorrectValue() {
         let another = PBXHeadersBuildPhase(reference: "ref",
-                                           buildActionMask: 0,
                                            files: Set(arrayLiteral: "333"),
+                                           buildActionMask: 0,
                                            runOnlyForDeploymentPostprocessing: 0)
         XCTAssertEqual(subject, another)
     }
